updated OptiX util. resolved warnings in common sources.

+++ utils/optixu_on_cudau.h
@@ -348,7 +348,8 @@ namespace optixu {
 
 template <>
 cudau::Buffer::operator optixu::BufferView() const {
-    return optixu::BufferView(getCUdeviceptr(), numElements(), stride());
+    return optixu::BufferView(
+        getCUdeviceptr(), static_cast<uint32_t>(numElements()), static_cast<uint32_t>(stride()));
 }
 
 //inline optixu::BufferView getView(const cudau::Buffer &buffer) {
